requirements.txt
# StudyOS — Python dependencies
# Install with: pip install -r requirements.txt

streamlit>=1.40.0
google-genai>=1.0.0
python-dotenv>=1.0.0
